[Code health] Move LatLngConverter from model to util package (#2522)
* Move LatLngConverter to util

* Rename method to toDmsFormat

* Convert to extension function

* Rename to CoordinatesExt

* Update license year and fix method name
